Update: Comment on stack vs. heap memory
I read how large numbers are better stored on heap over the stack. Storing a large enough integer on the stack can cause it to overflow. The stack is meant for fast quick operations with small to regular-sized data between functions. If large data is needed, it should be allocated to the heap to grow as large as needed.